+++ engine/src/main/java/org/terasology/rendering/nui/layers/mainMenu/advancedGameSetupScreen/DownloadPopupProgressListener.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.rendering.nui.layers.mainMenu.selectModulesScreen;
+package org.terasology.rendering.nui.layers.mainMenu.advancedGameSetupScreen;
 
 import org.terasology.rendering.nui.layers.mainMenu.WaitPopup;
 import org.terasology.utilities.download.MultiFileTransferProgressListener;
